+++ Source/VMware.vSphereDSC/VMware.vSphereDSC.psm1
@@ -536,6 +536,18 @@ class VMHostBaseDSC : BaseDSC {
     <#
     .DESCRIPTION
 
+    Specifies the time in minutes to wait for the VMHost to restart before timing out
+    and aborting the operation. The default value is 5 minutes.
+    #>
+    [DscProperty()]
+    [int] $RestartTimeoutMinutes = 5
+
+    hidden [string] $NotRespondingState = 'NotResponding'
+    hidden [string] $MaintenanceState = 'Maintenance'
+
+    <#
+    .DESCRIPTION
+
     Returns the VMHost with the specified Name on the specified Server.
     If the VMHost is not found, the method writes an error.
     #>
@@ -546,6 +558,98 @@ class VMHostBaseDSC : BaseDSC {
         }
         catch {
             throw "VMHost with name $($this.Name) was not found. For more information: $($_.Exception.Message)"
+        }
+    }
+
+    <#
+    .DESCRIPTION
+
+    Checks if the specified VMHost is in Maintenance mode and if not, throws an exception.
+    #>
+    [void] EnsureVMHostIsInMaintenanceMode($vmHost) {
+        if ($vmHost.ConnectionState.ToString() -ne $this.MaintenanceState) {
+            throw "The Resource Update operation requires the VMHost $($vmHost.Name) to be in a Maintenance mode."
+        }
+    }
+
+    <#
+    .DESCRIPTION
+
+    Ensures that the specified VMHost is restarted successfully in the specified period of time. If the elapsed time is
+    longer than the desired time for restart, the method throws an exception.
+    #>
+    [void] EnsureRestartTimeoutIsNotReached($elapsedTimeInSeconds) {
+        $timeSpan = New-TimeSpan -Seconds $elapsedTimeInSeconds
+        if ($this.RestartTimeoutMinutes -le $timeSpan.Minutes) {
+            throw "Aborting the operation. VMHost $($this.Name) could not be restarted successfully in $($this.RestartTimeoutMinutes) minutes."
+        }
+    }
+
+    <#
+    .DESCRIPTION
+
+    Ensures that the specified VMHost is in the desired state after successful restart operation.
+    #>
+    [void] EnsureVMHostIsInDesiredState($requiresVIServerConnection, $desiredState) {
+        $sleepTimeInSeconds = 10
+        $elapsedTimeInSeconds = 0
+
+        while ($true) {
+            $this.EnsureRestartTimeoutIsNotReached($elapsedTimeInSeconds)
+
+            Start-Sleep -Seconds $sleepTimeInSeconds
+            $elapsedTimeInSeconds += $sleepTimeInSeconds
+
+            try {
+                if ($requiresVIServerConnection) {
+                    $this.ConnectVIServer()
+                }
+
+                $vmHost = $this.GetVMHost()
+                if ($vmHost.ConnectionState.ToString() -eq $desiredState) {
+                    break
+                }
+
+                Write-Verbose -Message "VMHost $($this.Name) is still not in $desiredState State."
+            }
+            catch {
+                <#
+                Here the message used in the try block is written again in the case when an exception is thrown
+                when retrieving the VMHost or establishing a Connection. This way the user still gets notified
+                that the VMHost is not in the Desired State.
+                #>
+                Write-Verbose -Message "VMHost $($this.Name) is still not in $desiredState State."
+            }
+        }
+
+        Write-Verbose -Message "VMHost $($this.Name) is successfully restarted and in $desiredState State."
+    }
+
+    <#
+    .DESCRIPTION
+
+    Restarts the specified VMHost so that the Update of the VMHost Configuration is successful.
+    #>
+    [void] RestartVMHost($vmHost) {
+        try {
+            Restart-VMHost -Server $this.Connection -VMHost $vmHost -Confirm:$false -ErrorAction Stop
+        }
+        catch {
+            throw "Cannot restart VMHost $($vmHost.Name). For more information: $($_.Exception.Message)"
+        }
+
+        <#
+        If the Connection is directly to a vCenter we do not need to establish a new connection so we pass $false
+        to the method 'EnsureVMHostIsInCorrectState'. When the Connection is directly to an ESXi, after a successful
+        restart the ESXi is down so new Connection needs to be established to check the ESXi state. So we pass $true
+        to the method 'EnsureVMHostIsInCorrectState'.
+        #>
+        if ($this.Connection.ProductLine -eq $this.vCenterProductId) {
+            $this.EnsureVMHostIsInDesiredState($false, $this.NotRespondingState)
+            $this.EnsureVMHostIsInDesiredState($false, $this.MaintenanceState)
+        }
+        else {
+            $this.EnsureVMHostIsInDesiredState($true, $this.MaintenanceState)
         }
     }
 }
@@ -1959,7 +2063,7 @@ class VMHostAdvancedSettings : VMHostBaseDSC {
             if an invalid Advanced Setting is present in the passed hashtable and in the same time to
             provide an information to the user that invalid data is passed.
             #>
-            Write-Warning "Advanced Setting $advancedSettingName does not exist for VMHost $($vmHostName) and will be ignored."
+            Write-Warning -Message "Advanced Setting $advancedSettingName does not exist for VMHost $($vmHostName) and will be ignored."
         }
 
         return $advancedSetting
@@ -2671,6 +2775,126 @@ class VMHostNtpSettings : VMHostBaseDSC {
 
         $serviceSystem = Get-View -Server $this.Connection -Id $vmHost.ExtensionData.ConfigManager.ServiceSystem
         Update-ServicePolicy -ServiceSystem $serviceSystem -ServiceId $this.ServiceId -ServicePolicyValue $this.NtpServicePolicy.ToString().ToLower()
+    }
+}
+
+[DscResource()]
+class VMHostPciPassthrough : VMHostBaseDSC {
+    <#
+    .DESCRIPTION
+
+    Specifies the Id of the PCI Device, composed of "bus:slot.function".
+    #>
+    [DscProperty(Key)]
+    [string] $Id
+
+    <#
+    .DESCRIPTION
+
+    Specifies whether passThru has been configured for this device.
+    #>
+    [DscProperty(Mandatory)]
+    [bool] $Enabled
+
+    [void] Set() {
+    	$this.ConnectVIServer()
+        $vmHost = $this.GetVMHost()
+        $vmHostPciPassthruSystem = $this.GetVMHostPciPassthruSystem($vmHost)
+        $pciDevice = $this.GetPCIDevice($vmHostPciPassthruSystem)
+
+        $this.EnsurePCIDeviceIsPassthruCapable($pciDevice)
+        $this.EnsureVMHostIsInMaintenanceMode($vmHost)
+
+        $this.UpdatePciPassthruConfiguration($vmHostPciPassthruSystem)
+        $this.RestartVMHost($vmHost)
+    }
+
+    [bool] Test() {
+    	$this.ConnectVIServer()
+        $vmHost = $this.GetVMHost()
+        $vmHostPciPassthruSystem = $this.GetVMHostPciPassthruSystem($vmHost)
+
+        $pciDevice = $this.GetPCIDevice($vmHostPciPassthruSystem)
+        $this.EnsurePCIDeviceIsPassthruCapable($pciDevice)
+
+        return ($this.Enabled -eq $pciDevice.PassthruEnabled)
+    }
+
+    [VMHostPciPassthrough] Get() {
+        $result = [VMHostPciPassthrough]::new()
+        $result.Server = $this.Server
+
+    	$this.ConnectVIServer()
+        $vmHost = $this.GetVMHost()
+        $vmHostPciPassthruSystem = $this.GetVMHostPciPassthruSystem($vmHost)
+
+        $pciDevice = $this.GetPCIDevice($vmHostPciPassthruSystem)
+        $this.EnsurePCIDeviceIsPassthruCapable($pciDevice)
+
+        $result.Name = $vmHost.Name
+    	$result.Id = $pciDevice.Id
+        $result.Enabled = $pciDevice.PassthruEnabled
+
+        return $result
+    }
+
+    <#
+    .DESCRIPTION
+
+    Retrieves the PciPassthruSystem of the specified VMHost from the server.
+    #>
+    [PSObject] GetVMHostPciPassthruSystem($vmHost) {
+        try {
+            $vmHostPciPassthruSystem = Get-View -Server $this.Connection -Id $vmHost.ExtensionData.ConfigManager.PciPassthruSystem -ErrorAction Stop
+            return $vmHostPciPassthruSystem
+        }
+        catch {
+            throw "Could not retrieve the PciPassthruSystem of VMHost $($vmHost.Name). For more information: $($_.Exception.Message)"
+        }
+    }
+
+    <#
+    .DESCRIPTION
+
+    Retrieves the PCI Device with the specified Id from the server.
+    #>
+    [PSObject] GetPCIDevice($vmHostPciPassthruSystem) {
+        $pciDevice = $vmHostPciPassthruSystem.PciPassthruInfo | Where-Object { $_.Id -eq $this.Id }
+        if ($null -eq $pciDevice) {
+            throw "The specified PCI Device $($this.Id) does not exist for VMHost $($this.Name)."
+        }
+
+        return $pciDevice
+    }
+
+    <#
+    .DESCRIPTION
+
+    Checks if the specified PCIDevice is Passthrough capable and if not, throws an exception.
+    #>
+    [void] EnsurePCIDeviceIsPassthruCapable($pciDevice) {
+        if (!$pciDevice.PassthruCapable) {
+            throw "Cannot configure PCI-Passthrough on incapable device $($pciDevice.Id)."
+        }
+    }
+
+    <#
+    .DESCRIPTION
+
+    Performs an update on the specified PCI Device by changing its Passthru Enabled value.
+    #>
+    [void] UpdatePciPassthruConfiguration($vmHostPciPassthruSystem) {
+        $vmHostPciPassthruConfig = New-Object VMware.Vim.HostPciPassthruConfig
+
+        $vmHostPciPassthruConfig.Id = $this.Id
+        $vmHostPciPassthruConfig.PassthruEnabled = $this.Enabled
+
+        try {
+            Update-PassthruConfig -VMHostPciPassthruSystem $vmHostPciPassthruSystem -VMHostPciPassthruConfig $vmHostPciPassthruConfig
+        }
+        catch {
+            throw "The Update operation of PCI Device $($this.Id) failed with the following error: $($_.Exception.Message)"
+        }
     }
 }